+++ src/cadqueryhelper/__init__.py
@@ -4,3 +4,4 @@ from .Base import Base
 from .Hinge import Hinge
 from .HelperPlate import HelperPlate
 from .randomizedRotationGrid import randomized_rotation_grid
+from .schemeGrid import scheme_grid